+++ src/entities/create-form/index.tsx
@@ -1,0 +1,9 @@
+export { default as CheckBoxOption } from './ui/CheckBoxOption';
+export { default as DeleteButton } from './ui/DeleteButton';
+export { default as DropDownOption } from './ui/DropDownOption';
+export { default as FormTitle } from './ui/FormTitle';
+export { default as FormTypeSelect } from './ui/FormTypeSelect';
+export { default as MultipleChoicOption } from './ui/MultipleChoicOption';
+export { default as PictureOption } from './ui/PictureOption';
+export { default as RequiredToggle } from './ui/RequiredToggle';
+export { default as TextOption } from './ui/TextOption';
