+++ src/entities/create-form/ui/FormTitle/index.tsx
@@ -2,7 +2,9 @@ import React from 'react';
 
 const FormTitle = () => {
   return (
-    <input placeholder="제목 입력" className="w-full text-h4 text-black" />
+    <div className="w-[60%] border-b-1 border-solid border-gray-100 py-4">
+      <input placeholder="제목 입력" className="w-full text-h4 text-black" />
+    </div>
   );
 };
