+++ src/shared/types/create-form/type.ts
@@ -1,0 +1,7 @@
+import { ReactNode } from 'react';
+
+export interface Option {
+  value: string;
+  label: string;
+  icon: ReactNode;
+}
