+++ src/shared/ui/index.tsx
@@ -5,3 +5,4 @@ export { default as PageHeader } from './PageHeader';
 export { default as CheckBox } from './CheckBox';
 export { default as Loading } from './Loading';
 export { default as ToggleButton } from './ToggleButton';
+export { default as AddItemButton } from './AddItemButton';
